ci: Add extra build job with clang 19
clang 19 introduces new warnings (related to the use of the template
keyword) which seem to be firing a lot in ACTS. Because OneAPI 2025 is
built on clang 19, this is causing us problems when trying to compile
ACTS as a dependency of traccc. To alleviate this issue I would like to
introduce a clang 19-based build test so we can precent new warnings
from sneaking into the ACTS codebase.