build: Make third-party libraries system dependencies
ACTS currently supports downloading its dependencies via FetchContent
which has the problem that they are - by default - considered as files
belonging to the project itself. This is undesirable, as this may
introduce warnings into the build of ACTS that we cannot do anything
about. In CMake 3.25, FetchContent gained the `SYSTEM` flag which allows
dependencies fetched through it to be marked as system libraries, for
which no warnings are issued.

This commit turns all out third-party dependencies into system
dependencies.